didConfigResourcePromise
